requirements.txt
playwright==1.46.0
gspread==6.1.2
google-auth==2.34.0
